pyproject.toml
[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "openai>",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "tau2",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[tool.uv.sources]
tau2 = { git = "https://github.com/sierra-research/tau2-bench.git", branch = "main" }
